+++ charts-digital.go
@@ -34,7 +34,10 @@ func processDigitalData(writer http.ResponseWriter, workplaceName string, dateFr
 			db.Select("date_time, data").Where("date_time >= ?", dateFrom).Where("date_time <= ?", dateTo).Where("device_port_id = ?", port.ID).Order("date_time").Order("id").Find(&digitalData)
 			var portData PortData
 			portData.PortName = "ID" + strconv.Itoa(int(port.ID)) + ": " + port.Name + " (" + port.Unit + ")"
-			portData.PortColor = cachedDevicePortsColorsById[int(port.ID)]
+			portColor := cachedDevicePortsColorsById[int(port.ID)]
+			if portColor == "#000000" {
+				portData.PortColor = ""
+			}
 			if len(digitalData) > 0 {
 				var initialData Data
 				initialData.Time = dateFrom.Unix()