+++ CHANGELOG.md
@@ -14,6 +14,12 @@ to [Semantic Versioning](http://semver.org/spec/v2.0.0.html) rules.
 * ```Fixed``` for any bug fixes.
 * ```Security``` in case of vulnerabilities.
 
+## [2021.2.3.7] - 2021-06-07
+
+### Fixed
+
+- proper assigned user for downtimes data table II.
+
 ## [2021.2.3.3] - 2021-06-03
 
 ### Fixed